Move new code changes to right location
